+++ Pole.hpp
@@ -5,14 +5,14 @@
 class Pole {
     public:
     char stan;
-    int wielkosc;
+    float wielkosc;
     int posX;
     int posY;
     sf::Sprite spritePola;
     sf::Texture teksturaPola;
 
 
-    Pole(char stan, int wielkosc, int posX, int posY);
+    Pole(char stan, float wielkosc, int posX, int posY);
     Pole();
     void zmienStan(char znak);
     bool isEmpty();
